feat: add status colours to the ConnectionField Select
